requirements.txt
Flask==2.0.2
pandas==1.3.4
Pillow==8.4.0
torch==1.10.0
torchvision==0.11.1
Werkzeug==2.0.2
